requirements.txt
# Core dependencies
torch>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.26.0

# Vector stores
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment if you have CUDA GPU

# Data processing
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# Evaluation & Visualization  
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Development tools
pytest>=7.4.0
black>=23.7.0
